+++ test/services/onlinetesting/Phase1TestServiceSpec.scala
@@ -71,8 +71,8 @@ class Phase1TestServiceSpec extends PlaySpec with BeforeAndAfterEach with Mockit
   val bqScheduleId =testGatewayConfig.phase1Tests.scheduleIds("bq")
 
   val preferredName = "Preferred\tName"
-  val lastName = "Johnson"
   val preferredNameSanitized = "Preferred Name"
+  val lastName = "Jones"
   val userId = "userId"
 
   val onlineTestApplication = OnlineTestApplication(applicationId = "appId",
@@ -163,7 +163,7 @@ class Phase1TestServiceSpec extends PlaySpec with BeforeAndAfterEach with Mockit
       val result = phase1TestService.getTestProfile("valid-userid").futureValue
 
       result.get.expirationDate must equal(validExpireDate)
-//      result.get.activeTests.head.invitationDate must equal(InvitationDate)
+      //      result.get.activeTests.head.invitationDate must equal(InvitationDate)
     }
   }
 
@@ -231,52 +231,52 @@ class Phase1TestServiceSpec extends PlaySpec with BeforeAndAfterEach with Mockit
     }
     "fail, audit 'UserRegisteredForOnlineTest' and audit 'UserInvitedToOnlineTest' " +
       "if there is an exception retrieving the contact details" in new OnlineTest {
-        when(cubiksGatewayClientMock.registerApplicant(eqTo(registerApplicant))(any[HeaderCarrier]))
-          .thenReturn(Future.successful(registration))
-        when(cubiksGatewayClientMock.inviteApplicant(any[InviteApplicant])(any[HeaderCarrier]))
-          .thenReturn(Future.successful(invitation))
-        when(cdRepositoryMock.find(userId))
-          .thenReturn(Future.failed(new Exception))
+      when(cubiksGatewayClientMock.registerApplicant(eqTo(registerApplicant))(any[HeaderCarrier]))
+        .thenReturn(Future.successful(registration))
+      when(cubiksGatewayClientMock.inviteApplicant(any[InviteApplicant])(any[HeaderCarrier]))
+        .thenReturn(Future.successful(invitation))
+      when(cdRepositoryMock.find(userId))
+        .thenReturn(Future.failed(new Exception))
 
-        val result = phase1TestService.registerAndInviteForTestGroup(onlineTestApplication)
-        result.failed.futureValue mustBe an[Exception]
+      val result = phase1TestService.registerAndInviteForTestGroup(onlineTestApplication)
+      result.failed.futureValue mustBe an[Exception]
 
-        verify(auditServiceMock, times(2)).logEventNoRequest("UserRegisteredForOnlineTest", auditDetails)
-        verify(auditServiceMock, times(2)).logEventNoRequest("UserInvitedToOnlineTest", auditDetails)
-        verify(auditServiceMock, times(4)).logEventNoRequest(any[String], any[Map[String, String]])
-      }
+      verify(auditServiceMock, times(2)).logEventNoRequest("UserRegisteredForOnlineTest", auditDetails)
+      verify(auditServiceMock, times(2)).logEventNoRequest("UserInvitedToOnlineTest", auditDetails)
+      verify(auditServiceMock, times(4)).logEventNoRequest(any[String], any[Map[String, String]])
+    }
     "fail, audit 'UserRegisteredForOnlineTest' and audit 'UserInvitedToOnlineTest'" +
       " if there is an exception sending the invitation email" in new OnlineTest {
-        when(cubiksGatewayClientMock.registerApplicant(any[RegisterApplicant])(any[HeaderCarrier]))
-          .thenReturn(Future.successful(registration))
-        when(cubiksGatewayClientMock.inviteApplicant(any[InviteApplicant])(any[HeaderCarrier]))
-          .thenReturn(Future.successful(invitation))
-        when(cdRepositoryMock.find(userId))
-          .thenReturn(Future.successful(contactDetails))
+      when(cubiksGatewayClientMock.registerApplicant(any[RegisterApplicant])(any[HeaderCarrier]))
+        .thenReturn(Future.successful(registration))
+      when(cubiksGatewayClientMock.inviteApplicant(any[InviteApplicant])(any[HeaderCarrier]))
+        .thenReturn(Future.successful(invitation))
+      when(cdRepositoryMock.find(userId))
+        .thenReturn(Future.successful(contactDetails))
 
-        when(emailClientMock.sendOnlineTestInvitation(
-          eqTo(emailContactDetails), eqTo(preferredName), eqTo(expirationDate)
-        )(any[HeaderCarrier]))
-          .thenReturn(Future.failed(new Exception))
+      when(emailClientMock.sendOnlineTestInvitation(
+        eqTo(emailContactDetails), eqTo(preferredName), eqTo(expirationDate)
+      )(any[HeaderCarrier]))
+        .thenReturn(Future.failed(new Exception))
 
-        val result = phase1TestService.registerAndInviteForTestGroup(onlineTestApplication)
-        result.failed.futureValue mustBe an[Exception]
+      val result = phase1TestService.registerAndInviteForTestGroup(onlineTestApplication)
+      result.failed.futureValue mustBe an[Exception]
 
-        verify(auditServiceMock, times(2)).logEventNoRequest("UserRegisteredForOnlineTest", auditDetails)
-        verify(auditServiceMock, times(2)).logEventNoRequest("UserInvitedToOnlineTest", auditDetails)
-        verify(auditServiceMock, times(4)).logEventNoRequest(any[String], any[Map[String, String]])
-      }
+      verify(auditServiceMock, times(2)).logEventNoRequest("UserRegisteredForOnlineTest", auditDetails)
+      verify(auditServiceMock, times(2)).logEventNoRequest("UserInvitedToOnlineTest", auditDetails)
+      verify(auditServiceMock, times(4)).logEventNoRequest(any[String], any[Map[String, String]])
+    }
     "fail, audit 'UserRegisteredForOnlineTest', audit 'UserInvitedToOnlineTest'" +
       ", not send invitation email to user" +
       "if there is an exception storing the status and the online profile data to database" in new OnlineTest {
-        when(cubiksGatewayClientMock.registerApplicant(eqTo(registerApplicant))(any[HeaderCarrier]))
-          .thenReturn(Future.successful(registration))
-        when(cubiksGatewayClientMock.inviteApplicant(any[InviteApplicant])(any[HeaderCarrier]))
-          .thenReturn(Future.successful(invitation))
-        when(cdRepositoryMock.find(userId)).thenReturn(Future.successful(contactDetails))
-        when(emailClientMock.sendOnlineTestInvitation(
-          eqTo(emailContactDetails), eqTo(preferredName), eqTo(expirationDate))(any[HeaderCarrier])
-        ).thenReturn(Future.successful(()))
+      when(cubiksGatewayClientMock.registerApplicant(eqTo(registerApplicant))(any[HeaderCarrier]))
+        .thenReturn(Future.successful(registration))
+      when(cubiksGatewayClientMock.inviteApplicant(any[InviteApplicant])(any[HeaderCarrier]))
+        .thenReturn(Future.successful(invitation))
+      when(cdRepositoryMock.find(userId)).thenReturn(Future.successful(contactDetails))
+      when(emailClientMock.sendOnlineTestInvitation(
+        eqTo(emailContactDetails), eqTo(preferredName), eqTo(expirationDate))(any[HeaderCarrier])
+      ).thenReturn(Future.successful(()))
 
 
       when(otRepositoryMock.insertOrUpdateTestGroup("appId", phase1TestProfile))
@@ -420,7 +420,7 @@ class Phase1TestServiceSpec extends PlaySpec with BeforeAndAfterEach with Mockit
       when(otRepositoryMock.insertOrUpdateTestGroup(any[String], any[Phase1TestProfile]))
         .thenReturn(Future.successful(()))
       when(otRepositoryMock.updateProgressStatus(any[String], any[ProgressStatus]))
-          .thenReturn(Future.successful(()))
+        .thenReturn(Future.successful(()))
 
       val result = phase1TestService.markAsReportReadyToDownload(cubiksUserId, reportReady).futureValue
 
@@ -441,7 +441,7 @@ class Phase1TestServiceSpec extends PlaySpec with BeforeAndAfterEach with Mockit
       when(otRepositoryMock.insertOrUpdateTestGroup(any[String], any[Phase1TestProfile]))
         .thenReturn(Future.successful(()))
       when(otRepositoryMock.updateProgressStatus(any[String], any[ProgressStatus]))
-          .thenReturn(Future.successful(()))
+        .thenReturn(Future.successful(()))
 
       val result = phase1TestService.markAsReportReadyToDownload(cubiksUserId, reportReady).futureValue
 
@@ -475,16 +475,16 @@ class Phase1TestServiceSpec extends PlaySpec with BeforeAndAfterEach with Mockit
 
   "retrieve phase 1 test report" should {
     "return an exception if no report Id is set" in new OnlineTest {
-        an[Exception] must be thrownBy phase1TestService.retrieveTestResult(Phase1TestWithUserIds(
-          "appId", "userId", phase1TestProfile
-        ))
+      an[Exception] must be thrownBy phase1TestService.retrieveTestResult(Phase1TestWithUserIds(
+        "appId", "userId", phase1TestProfile
+      ))
     }
 
     "return an exception if there is an error retrieving one of the reports" in new OnlineTest {
       val failedTest = phase1Test.copy(scheduleId = 555, reportId = Some(2))
       val successfulTest = phase1Test.copy(scheduleId = 444, reportId = Some(1))
 
-       when(cubiksGatewayClientMock.downloadXmlReport(eqTo(successfulTest.reportId.get))(any[HeaderCarrier]))
+      when(cubiksGatewayClientMock.downloadXmlReport(eqTo(successfulTest.reportId.get))(any[HeaderCarrier]))
         .thenReturn(Future.successful(OnlineTestCommands.TestResult(status = "Completed",
           norm = "some norm",
           tScore = Some(23.9999d),
@@ -557,8 +557,6 @@ class Phase1TestServiceSpec extends PlaySpec with BeforeAndAfterEach with Mockit
       val actor = ActorSystem()
     }
   }
-
-
 
   trait SuccessfulTestInviteFixture extends OnlineTest {
     when(cubiksGatewayClientMock.registerApplicant(eqTo(registerApplicant))(any[HeaderCarrier]))
